Template fragment cache extension
The extension caches everything within its block in APC.

The blocks are as below:
{% cache %}
....
{% endcache %}

or

{% cache 'my_custom_cache_key' %}
....
{% endcache %}

or

{% cache 'my_custom_cache_key_' ~ variable %}
....
{% endcache %}

They both save the entire contents of the block, and if you refresh a page,
or if same block is repeated somewhere else the cached content will be outputed
instead of having to execute everything within again.

The block without a give cache key, automatically generates a cache key using the filename + lineno.