Use DateTimeZone object instead of the name in localizeddate filter
In my case, my timezone's name returned by ``DateTime`` object is "+02:00". When I use this name with the ``localizedate()`` function, it correctly convert the date with the ``twig_date_converter()`` but the ``IntlDateFormater::create()`` returns ``null``.

When I read the doc and this issue : https://bugs.php.net/bug.php?id=66323, I have concluded that my name doesn't match with the ICU database as needed.

As to give a ``DateTimeZone`` object is possible, I think it could be more interesting to give the date's timezone object instead of it's name to avoid this kind of errors.